Start debug day1 server.

+++ day1-codec/codec/codec.go
@@ -4,7 +4,7 @@ import "io"
 
 type Header struct {
 	ServiceMethod string
-	Seq           int
+	Seq           uint64
 	Error         string
 }
